+++ src/GalleryImage.js
@@ -4,7 +4,7 @@ import CircularProgress from '@material-ui/core/CircularProgress';
 import Typography from '@material-ui/core/Typography';
 import React from 'react';
 import {drawEmbeddingImage, getSpotRadius} from './ImageChart';
-import {drawLabels, getVisualizer, setAxesColors} from './ScatterChartThree';
+import {drawLabels, getVisualizer} from './ScatterChartThree';
 import {
     getCategoryLabelsPositions,
     getLabels,
@@ -34,14 +34,15 @@ class GalleryImage extends React.PureComponent {
             scatterPlot,
             selection,
             traceInfo,
-            unselectedMarkerOpacity
+            unselectedMarkerOpacity,
+            unselectedPointSize
         } = this.props;
 
         if (traceInfo.type === 'scatter') {
             let spriteVisualizer = getVisualizer(scatterPlot, POINT_VISUALIZER_ID);
             spriteVisualizer.zoomFactor = this.zoomFactor;
 
-            updateScatterChart(scatterPlot, traceInfo, selection, markerOpacity, unselectedMarkerOpacity, pointSize,
+            updateScatterChart(scatterPlot, traceInfo, selection, markerOpacity, unselectedMarkerOpacity, pointSize, unselectedPointSize,
                 categoricalNames, chartOptions, obsCat, cachedData, traceInfo.camera);
 
             const canvas = containerElement.querySelector('canvas');