+++ src/ThreeUtil.js
@@ -1,4 +1,4 @@
-import {makeStyles, ScatterPlot, ScatterPlotVisualizerSprites, ScatterPlotVisualizerSvgLabels} from 'scatter-gl';
+import {makeStyles, ScatterPlot, ScatterPlotVisualizerSprites, ScatterPlotVisualizerSvgLabels} from './scatter-gl';
 import {Color, OrthographicCamera, Vector3} from 'three';
 import {getEmbeddingKey} from './actions';
 import {getVisualizer, setAxesColors} from './ScatterChartThree';
@@ -250,7 +250,7 @@ export function getCategoryLabelsPositions(embedding, obsKeys, cachedData) {
     return {labels: labelValues, positions: labelPositions};
 }
 
-export function updateScatterChart(scatterPlot, traceInfo, selection, markerOpacity, unselectedMarkerOpacity, pointSize, categoricalNames = {}, chartOptions, obsCatKeys, cachedData, cameraDef) {
+export function updateScatterChart(scatterPlot, traceInfo, selection, markerOpacity, unselectedMarkerOpacity, pointSize, unselectedPointSize, categoricalNames = {}, chartOptions, obsCatKeys, cachedData, cameraDef) {
     const is3d = traceInfo.z != null;
     const colors = traceInfo.colors;
     let positions = traceInfo.positions;
@@ -261,10 +261,14 @@ export function updateScatterChart(scatterPlot, traceInfo, selection, markerOpac
     if (updateZ) {
         positions = positions.slice();
     }
-
+    const scale = new Float32Array(traceInfo.x.length);
+    scale.fill(pointSize);
     for (let i = 0, j = 3, k = 2; i < npoints; i++, j += 4, k += 3) {
         const isSelected = isSelectionEmpty || selection.has(i);
         colors[j] = isSelected ? markerOpacity : unselectedMarkerOpacity;
+        if (!isSelected) {
+            scale[i] = unselectedPointSize;
+        }
         if (updateZ && !isSelected) {
             positions[k] = -1;
         }
@@ -284,16 +288,15 @@ export function updateScatterChart(scatterPlot, traceInfo, selection, markerOpac
     scatterPlot.setPointColors(colors);
     scatterPlot.setPointPositions(positions);
 
-    const scale = new Float32Array(traceInfo.x.length);
-    scale.fill(pointSize);
-    if (!isSelectionEmpty && is3d && unselectedMarkerOpacity === 0) { // hide filtered points so they don't obscure non-filtered points
-        for (let i = 0, j = 3, k = 2; i < npoints; i++, j += 4, k += 3) {
-            if (colors[j] === 0) {
-                scale[i] = 0;
-            }
-        }
 
-    }
+    // if (!isSelectionEmpty && is3d && unselectedMarkerOpacity === 0) { // hide filtered points so they don't obscure non-filtered points
+    //     for (let i = 0, j = 3, k = 2; i < npoints; i++, j += 4, k += 3) {
+    //         if (colors[j] === 0) {
+    //             scale[i] = 0;
+    //         }
+    //     }
+    //
+    // }
     scatterPlot.setPointScaleFactors(scale);
 
     const showLabels = obsCatKeys.length > 0;